refactor: split out shared utils
